Archives/mcp-servers-v1/web-agent/pyproject.toml
[project]
name = "web-agent"
version = "0.1.0"
description = "WebAgent — intelligent web browsing, scraping, search via Agent SDK"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.42",
    "fastmcp>=2.0.0",
    "python-dotenv>=1.0",
    "playwright>=1.49",
    "httpx>=0.27",
]
